@@ -1,13 +1,24 @@
-plugins {
-    id "org.springframework.boot" version '2.4.1'
-    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
-    id 'java'
+buildscript {
+    ext {
+        springBootVersion = '2.3.4.RELEASE'
+
+    }
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
+    }
 }
+
+apply plugin: 'java'
+apply plugin: 'org.springframework.boot'
+apply plugin: 'io.spring.dependency-management'
 
 
 group = 'com.reeco'
 version = '0.1.0'
-sourceCompatibility = "1.9"
+sourceCompatibility = "11"
 
 
 ext {
@@ -27,32 +38,29 @@ repositories {
 
 
 dependencies {
-
+    implementation project(':common')
     implementation 'org.jetbrains:annotations:20.1.0'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
-
-    // implementation('com.marvin.shares.api:api')
-    // implementation('com.marvin.shares.util:util')
-
-
-//    implementation('org.springframework.boot:spring-boot-starter-actuator')
+    implementation('org.springframework.boot:spring-boot-starter-actuator')
     implementation ('org.springframework.boot:spring-boot-starter-web')
 //    implementation('io.springfox:springfox-boot-starter:3.0.0')
     implementation('org.springframework.boot:spring-boot-starter-data-cassandra:2.5.4')
     implementation ('org.springframework.boot:spring-boot-starter-parent:2.1.12.RELEASE')
     implementation 'com.fasterxml.jackson.core:jackson-databind'
-//    implementation ('org.springframework.boot:spring-boot-starter-data-jdbc')
-//    implementation ('org.springframework.boot:spring-boot-starter-data-jpa')
-//    implementation ('org.springframework.boot:spring-boot-starter-jdbc')
     implementation('org.postgresql:postgresql')
     implementation "org.apache.commons:commons-csv:1.8"
-    // implementation('org.springframework.boot:spring-boot-starter-data-mongodb-reactive')
 
 
     testImplementation('org.springframework.boot:spring-boot-starter-test')
     testImplementation('io.projectreactor:reactor-test')
+
+    implementation("org.mapstruct:mapstruct:${mapstructVersion}")
+    compileOnly "org.mapstruct:mapstruct-processor:${mapstructVersion}"
+    annotationProcessor "org.mapstruct:mapstruct-processor:${mapstructVersion}"
+    testAnnotationProcessor "org.mapstruct:mapstruct-processor:${mapstructVersion}"
 }
 
-
-targetCompatibility = JavaVersion.VERSION_1_9
+test {
+    useJUnitPlatform()
+}
