@@ -1,7 +1,7 @@
 package com.reeco.common.model.enumtype;
 
 public enum FrequenceType {
-    S(0.166667), M(1d), H(60d), D(1440d);
+    SECOND(0.166667), MINUTE(1d), HOUR(60d), DAY(1440d);
 
     private Double value;
 
