@@ -20,13 +20,16 @@ public interface NumericalTsByOrgRepository extends CassandraRepository<Numerica
     @Query("select * from numeric_series_by_organization\n" +
             "where event_time >= ?0 and event_time <= ?1 \n" +
             "  and organization_id = ?2  \n" +
-            "  and param_id = ?3 \n" +
-            "  ALLOW FILTERING ;")
+            "  and param_id = ?3;")
     List<NumericalTsByOrg> findDataDetail(Timestamp startTime, Timestamp endTime, Long organizationId, Long paramId);
 
     @Query("SELECT * FROM numeric_series_by_organization WHERE organization_id=?0 and param_id=?1\n" +
-            "LIMIT 2 ALLOW FILTERING ;")
+            "LIMIT 2;")
     List<NumericalTsByOrg> find2LatestRow(Long orgId, Long paramId);
+
+    @Query("SELECT * FROM numeric_series_by_organization WHERE organization_id=?0 and param_id=?1\n" +
+            "LIMIT 1;")
+    Optional<NumericalTsByOrg> find1LatestRow(Long orgId, Long paramId);
 
     Optional<NumericalTsByOrg> findByPartitionKey(NumericalTsByOrg.Key key);
 }
