@@ -1,9 +1,8 @@
 package com.reeco.ingestion.application.port.in;
 
-import com.reeco.ingestion.domain.Parameter;
-import com.reeco.ingestion.domain.ParamsByOrg;
+import com.reeco.ingestion.domain.ParamsAndOrg;
 import reactor.core.publisher.Flux;
 
 public interface LoadOrgAndParamPort {
-    Flux<ParamsByOrg> findAllParamsGroupByOrg();
+    Flux<ParamsAndOrg> findAllParamsGroupByOrg();
 }
