@@ -4,10 +4,11 @@ import com.reeco.ingestion.domain.NumericalStatEvent;
 import reactor.core.publisher.Flux;
 
 import java.sql.Timestamp;
+import java.time.LocalDateTime;
 
 public interface AggregateEventsPort {
 
-    Flux<NumericalStatEvent> aggEventByOrgAndParams(Timestamp startTime, Timestamp endTime);
+    Flux<NumericalStatEvent> aggEventByOrgAndParams(LocalDateTime endTime);
 
 
 }
