@@ -1,8 +1,9 @@
 package com.reeco.ingestion.application.port.out;
 
 import java.sql.Timestamp;
+import java.time.LocalDateTime;
 
 public interface NumStatRepository {
-    void insert(Timestamp startTime, Timestamp endTime);
+    void insert(LocalDateTime endTime);
 
 }
