@@ -2,16 +2,14 @@ package com.reeco.ingestion.infrastructure.persistence.cassandra.entity;
 
 import com.reeco.common.framework.EnumNamePattern;
 import com.reeco.common.model.enumtype.ParamType;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.Getter;
-import lombok.ToString;
+import lombok.*;
 import org.springframework.data.cassandra.core.cql.PrimaryKeyType;
 import org.springframework.data.cassandra.core.mapping.*;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
+import java.time.LocalDateTime;
 
 @Table("params_by_organization")
 @AllArgsConstructor
@@ -71,4 +69,8 @@ public class ParamsByOrg {
 
 //    @Column("updated_at")
 //    private LocalDateTime updatedAt;
+
+    @Column("last_agg_time")
+    @Setter
+    private LocalDateTime lastAggTime;
 }
